Adjust to 'domain's current API
'domain' uses 'SignError' for representing on-disk cryptography
failures.  This is not appropriate for KMIP, but currently we implement
the same 'SignRaw' trait anyway.  This means we can't stuff enough
error information in the 'SignError' type.  We should stop implementing
'SignRaw', use a custom error type, and provide an asynchronous API.